+++ examples/beep.rs
@@ -28,8 +28,6 @@ where
     let channels = config.channels as usize;
 
     //let c = mls();
-    //let c = mls() >> lowpole_hz(400.0) >> lowpole_hz(400.0);
-    //let c = (mls() | dc(500.0)) >> butterpass();
     //let c = (mls() | dc(400.0) | dc(50.0)) >> resonator();
     //let c = pink();
 
@@ -39,11 +37,11 @@ where
     //let c = oversample(sine_hz(f) * f * m + f >> sine());
 
     // Pulse wave.
-    let c = lfo(|t| {
-        let pitch = 110.0;
-        let duty = lerp11(0.01, 0.99, sin_hz(0.05, t));
-        (pitch, duty)
-    }) >> pulse();
+    //let c = lfo(|t| {
+    //    let pitch = 220.0;
+    //    let duty = lerp11(0.01, 0.99, sin_hz(0.05, t));
+    //    (pitch, duty)
+    //}) >> pulse();
 
     //let c = zero() >> pluck(220.0, 0.8, 0.8);
     //let c = dc(110.0) >> dsf_saw_r(0.99);
@@ -51,40 +49,38 @@ where
     //let c = dc(110.0) >> soft_saw();
     //let c = lfo(|t| xerp11(20.0, 2000.0, sin_hz(0.1, t))) >> dsf_square_r(0.99) >> lowpole_hz(1000.0);
     //let c = dc(110.0) >> square();
-    //let c = 0.2 * (organ_hz(midi_hz(57.0)) + organ_hz(midi_hz(61.0)) + organ_hz(midi_hz(64.0)));
-    //let c = organ_hz(110.0);
+    let c = 0.2 * (organ_hz(midi_hz(57.0)) + organ_hz(midi_hz(61.0)) + organ_hz(midi_hz(64.0)));
+    //let c = organ_hz(110.1) + organ_hz(54.9);
 
     // Filtered noise tone.
-    //let c = noise() >> resonator_hz(440.0, 5.0);
+    //let c = (noise() | dc((440.0, 50.0))) >> !resonator() >> resonator();
 
     // Test ease_noise.
     //let c = lfo(|t| xerp11(50.0, 5000.0, ease_noise(smooth9, 0, t))) >> triangle();
 
     // Bandpass filtering.
-    //let c = c
-    //    >> (pass() | envelope(|t| xerp(500.0, 20000.0, sin_hz(0.0666, t))))
-    //    >> bandpass_q(1.0);
+    let c = c >> (pass() | envelope(|t| xerp11(500.0, 5000.0, sin_hz(0.05, t)))) >> bandpass_q(5.0);
 
-    // Waveshapers.
-    //let c = c >> shape_fn(|x| tanh(x * 5.0));
+    // Waveshaper.
+    //let c = c >> shape(Shape::Crush(20.0));
 
     // Add feedback delay.
-    //let c = c & c >> feedback(butterpass_hz(1000.0) >> delay(1.0) * 0.5);
+    //let c = c >> (pass() & feedback(butterpass_hz(1000.0) >> delay(1.0) * 0.5));
 
     // Apply Moog filter.
-    //let c = (c | lfo(|t| (xerp11(110.0, 11000.0, sin_hz(0.15, t)), 0.6))) >> moog();
+    //let c = (c | lfo(|t| (xerp11(110.0, 11000.0, sin_hz(0.1, t)), 0.6))) >> moog();
 
-    let c = c >> split::<U2>();
+    let c = c >> pan(0.0);
 
     //let c = fundsp::sound::risset_glissando(false);
 
     // Add chorus.
-    //let c = c >> (chorus(0, 0.0, 0.01, 0.5) | chorus(1, 0.0, 0.01, 0.5));
+    let c = c >> (chorus(0, 0.0, 0.01, 0.2) | chorus(1, 0.0, 0.01, 0.2));
 
     // Add flanger.
-    let c = c
-        >> (flanger(0.6, 0.005, 0.01, |t| lerp11(0.005, 0.01, sin_hz(0.1, t)))
-            | flanger(0.6, 0.005, 0.01, |t| lerp11(0.005, 0.01, cos_hz(0.1, t))));
+    //let c = c
+    //    >> (flanger(0.6, 0.005, 0.01, |t| lerp11(0.005, 0.01, sin_hz(0.1, t)))
+    //        | flanger(0.6, 0.005, 0.01, |t| lerp11(0.005, 0.01, cos_hz(0.1, t))));
 
     // Add phaser.
     //let c = c