+++ src/hacker.rs
@@ -621,13 +621,19 @@ where
 /// - Input 0: x
 /// - Output(s): envelope linearly interpolated from samples at 2 ms intervals (average).
 #[inline]
-pub fn envelope2<E, R>(f: E) -> An<Envelope2<f64, f64, E, R>>
+pub fn envelope2<E, R>(
+    f: E,
+) -> An<EnvelopeIn<f64, f64, impl Fn(f64, &Frame<f64, U1>) -> R + Sized + Clone, U1, R>>
 where
     E: Fn(f64, f64) -> R + Clone,
     R: ConstantFrame<Sample = f64>,
     R::Size: Size<f64>,
 {
-    An(Envelope2::new(0.002, DEFAULT_SR, f))
+    An(EnvelopeIn::new(
+        0.002,
+        DEFAULT_SR,
+        move |t, i: &Frame<f64, U1>| f(t, i[0]),
+    ))
 }
 
 /// Control envelope from time-varying, input dependent function `f(t, x)` with `t` in seconds.
@@ -636,13 +642,19 @@ where
 /// - Input 0: x
 /// - Output(s): envelope linearly interpolated from samples at 2 ms intervals (average).
 #[inline]
-pub fn lfo2<E, R>(f: E) -> An<Envelope2<f64, f64, E, R>>
+pub fn lfo2<E, R>(
+    f: E,
+) -> An<EnvelopeIn<f64, f64, impl Fn(f64, &Frame<f64, U1>) -> R + Sized + Clone, U1, R>>
 where
     E: Fn(f64, f64) -> R + Clone,
     R: ConstantFrame<Sample = f64>,
     R::Size: Size<f64>,
 {
-    An(Envelope2::new(0.002, DEFAULT_SR, f))
+    An(EnvelopeIn::new(
+        0.002,
+        DEFAULT_SR,
+        move |t, i: &Frame<f64, U1>| f(t, i[0]),
+    ))
 }
 
 /// Control envelope from time-varying, input dependent function `f(t, x, y)` with `t` in seconds.
@@ -652,13 +664,19 @@ where
 /// - Input 1: y
 /// - Output(s): envelope linearly interpolated from samples at 2 ms intervals (average).
 #[inline]
-pub fn envelope3<E, R>(f: E) -> An<Envelope3<f64, f64, E, R>>
+pub fn envelope3<E, R>(
+    f: E,
+) -> An<EnvelopeIn<f64, f64, impl Fn(f64, &Frame<f64, U2>) -> R + Sized + Clone, U2, R>>
 where
     E: Fn(f64, f64, f64) -> R + Clone,
     R: ConstantFrame<Sample = f64>,
     R::Size: Size<f64>,
 {
-    An(Envelope3::new(0.002, DEFAULT_SR, f))
+    An(EnvelopeIn::new(
+        0.002,
+        DEFAULT_SR,
+        move |t, i: &Frame<f64, U2>| f(t, i[0], i[1]),
+    ))
 }
 
 /// Control envelope from time-varying, input dependent function `f(t, x, y)` with `t` in seconds.
@@ -668,13 +686,53 @@ where
 /// - Input 1: y
 /// - Output(s): envelope linearly interpolated from samples at 2 ms intervals (average).
 #[inline]
-pub fn lfo3<E, R>(f: E) -> An<Envelope3<f64, f64, E, R>>
+pub fn lfo3<E, R>(
+    f: E,
+) -> An<EnvelopeIn<f64, f64, impl Fn(f64, &Frame<f64, U2>) -> R + Sized + Clone, U2, R>>
 where
     E: Fn(f64, f64, f64) -> R + Clone,
     R: ConstantFrame<Sample = f64>,
     R::Size: Size<f64>,
 {
-    An(Envelope3::new(0.002, DEFAULT_SR, f))
+    An(EnvelopeIn::new(
+        0.002,
+        DEFAULT_SR,
+        move |t, i: &Frame<f64, U2>| f(t, i[0], i[1]),
+    ))
+}
+
+/// Control envelope from time-varying, input dependent function `f(t, i)` with `t` in seconds
+/// and `i` of type `&Frame<f64, I>` where `I` is the number of input channels.
+/// Spaces samples using pseudorandom jittering.
+/// Synonymous with `lfo_in`.
+/// - Inputs: i
+/// - Output(s): envelope linearly interpolated from samples at 2 ms intervals (average).
+pub fn envelope_in<E, I, R>(f: E) -> An<EnvelopeIn<f64, f64, E, I, R>>
+where
+    E: Fn(f64, &Frame<f64, I>) -> R + Clone,
+    I: Size<f64>,
+    R: ConstantFrame<Sample = f64>,
+    R::Size: Size<f64>,
+    R::Size: Size<f64>,
+{
+    An(EnvelopeIn::new(0.002, DEFAULT_SR, f))
+}
+
+/// Control envelope from time-varying, input dependent function `f(t, i)` with `t` in seconds
+/// and `i` of type `&Frame<f64, I>` where `I` is the number of input channels.
+/// Spaces samples using pseudorandom jittering.
+/// Synonymous with `envelope_in`.
+/// - Inputs: i
+/// - Output(s): envelope linearly interpolated from samples at 2 ms intervals (average).
+pub fn lfo_in<E, I, R>(f: E) -> An<EnvelopeIn<f64, f64, E, I, R>>
+where
+    E: Fn(f64, &Frame<f64, I>) -> R + Clone,
+    I: Size<f64>,
+    R: ConstantFrame<Sample = f64>,
+    R::Size: Size<f64>,
+    R::Size: Size<f64>,
+{
+    An(EnvelopeIn::new(0.002, DEFAULT_SR, f))
 }
 
 /// ADSR envelope.
@@ -696,7 +754,7 @@ pub fn adsr_live(
     decay: f64,
     sustain: f64,
     release: f64,
-) -> An<Envelope2<f64, f64, impl Fn(f64, f64) -> f64 + Sized + Clone, f64>> {
+) -> An<EnvelopeIn<f64, f64, impl Fn(f64, &Frame<f64, U1>) -> f64 + Sized + Clone, U1, f64>> {
     super::adsr::adsr_live(attack, decay, sustain, release)
 }