+++ CHANGES.md
@@ -5,6 +5,7 @@
 - Composable setting system for making it easier to control basic settings in real time.
 - `system` opcode was renamed `update`.
 - New waveforms, organ and soft saw.
+- New envelope opcodes with any number of inputs: `envelope_in` and `lfo_in`.
 
 ### Version 0.10